Added chat functionality to work in real-time. Changed nav menu styling and functionality for mobile site.

+++ tracker/models.py
@@ -64,7 +64,7 @@ class Exercise(models.Model):
     user = models.ForeignKey(User, on_delete=models.CASCADE, null=True)
     name = models.CharField(max_length=50, null=True)
     type = models.CharField(max_length=20, choices=TYPE, null=True)
-    calspermin = models.IntegerField(null=True)
+    calspermin = models.DecimalField(max_digits=30, decimal_places=15, null=True)
 
     def __str__(self):
         return f'Exercise {self.name}'
